pyproject.toml
[project]
name = "autoresearcher-litter-detection"
version = "0.1.0"
description = "Autonomous CNN training for litter segmentation"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "mlflow>=2.13.0",
    "datasets>=2.19.0",
    "Pillow>=10.3.0",
    "numpy>=1.26.0",
    "matplotlib>=3.9.0",
    "tqdm>=4.66.0",
    "scikit-learn>=1.5.0",
    "pycocotools>=2.0.8",
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
    "ipywidgets>=8.1.0",
    "huggingface-hub>=0.23.0",
    "albumentations>=1.4.0",
    "opencv-python-headless>=4.10.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-exporter-otlp>=1.21.0",
    "opentelemetry-instrumentation>=0.42b0",
    "eclipse-zenoh>=1.9.0",
    "ultralytics>=8.2.0",
    "ty>=0.0.29",
    "pre-commit>=4.6.0",
    "onnx>=1.16.0",
    "msgspec>=0.21.1",
    "pre-commit>=4.6.0",
    "onnx>=1.16.0",
    "msgspec>=0.21.1",
    "loguru>=0.7.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=8.2.0",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }

[project.optional-dependencies]
cpu = ["onnxruntime>=1.18.0"] # uv sync --extra cpu
gpu = ["onnxruntime-gpu>=1.18.0"] # uv sync --extra gpu
robot = [
    "unitree-webrtc-connect",
    "aiortc",
    "av",
] # uv sync --extra robot — only for real Go2 hardware tests

[project.scripts]
orch = "Orchastrator:main"
